Add plugins outside 1/.vim/plugged area.
Add rbenv, ruby-build and rbenv-gem-rehash github repos to manage
them with vim-plug. And replace fzf to see all outside plugins in
one place.
I add  an if section that it could be downloaded them if the system
is linux, because I use windows system as well, but I don't use these
plugins in windows only in linux system. And I don't know how I can
use them in windows to work good.